AV1e HWA VPL fixes (#1963)
* [AV1e HWA] Rename genx isa files from tgllp to gen12lp

Fei plugin cpp removed.

Issue: VSMGWL-38793
Test: manual

* [AV1e HWA] KW and warnings fixes

Issue: VSMGWL-38793, VSMGWL-39706, VSMGWL-39577
Test: manual

* [AV1e HWA] Review fixes

Issue: VSMGWL-38793, VSMGWL-39706, VSMGWL-39577
Test: manual

+++ _studio/mfx_lib/optimization/av1/avx2/mfx_av1_fwd_transform_avx2.cpp
@@ -3314,7 +3314,7 @@ namespace AV1PP {
         const int txw = tx_size_wide[size];
         for (int i = 0; i < txh; i++)
             for (int j = 0; j < txw; j++)
-                dst[i * txw + j] = dst32[i * txw + j];
+                dst[i * txw + j] = (short) dst32[i * txw + j];
     }
     template void ftransform_av1_avx2<TX_4X4,      DCT_DCT>(const short*,short*,int);
     template void ftransform_av1_avx2<TX_4X4,     ADST_DCT>(const short*,short*,int);
